// ColorPicker/src/constants/enums.ts
export enum EventType {
    PICKER,
    SLIDER,
    PANEL,
    HEX,
    RGB
}

export const destinations: string[] = [
    'ColorPicker',
    'ColorSlider',
    'ColorPanel',
    'ColorFields',
    'ColorLabel',
    'ColorSwatch'
]

export enum Tag {
    ANCHOR = "a",
    BODY = "body",
    BUTTON = "button",
    CANVAS = "canvas",
    DIV = "div",
    H1 = "h1",
    H2 = "h2",
    H3 = "h3",
    HEAD = "head",
    HTML = "html",
    INPUT = "input",
    LINK = "link",
    PARA = "p",
    SPAN = "span",
    STYLE = "style",
    TITLE = "title",
}

export enum Colors {
    RED = "#FF0000",
    YELLOW = "#FFFF00",
    GREEN = "#00FF00",
    CYAN = "#00FFFF",
    BLUE = "#0000FF",
    MAGENTA = "#FF00FF",
    WHITE = "#FFFFFF",
    BLACK = "#000000"
}